+++ src/components/[guild]/SetVisibility/hooks/useVisibilityTooltipLabel.tsx
@@ -1,0 +1,46 @@
+import RoleTag from "components/[guild]/RoleTag"
+import useGuild from "components/[guild]/hooks/useGuild"
+import { ReactNode } from "react"
+import { Visibility } from "types"
+import { VISIBILITY_DATA } from "../visibilityData"
+
+const useVisibilityTooltipLabel = (
+  visibility: Visibility,
+  visibilityRoleId: number
+): ReactNode => {
+  const { roles } = useGuild()
+  const role =
+    visibilityRoleId && visibility === Visibility.PRIVATE
+      ? roles?.find(({ id }) => id === visibilityRoleId)
+      : undefined
+
+  const tooltipDescription =
+    visibility !== Visibility.PRIVATE ? (
+      VISIBILITY_DATA[visibility].description
+    ) : !visibilityRoleId ? (
+      "Only visible to role holders"
+    ) : (
+      <>
+        Only visible to users that satisfy the
+        {role ? (
+          <RoleTag
+            size={"sm"}
+            bgColor={"blackAlpha.600"}
+            name={role.name}
+            imageUrl={role.imageUrl}
+          />
+        ) : (
+          "selected"
+        )}{" "}
+        role
+      </>
+    )
+
+  return (
+    <>
+      {VISIBILITY_DATA[visibility].title}: {tooltipDescription}
+    </>
+  )
+}
+
+export default useVisibilityTooltipLabel
